INTERLOK-3765 #comment Revert back the gradle version used by the installer.
The reason is that gradle 6.x changed the way it works with optional depedencies.
It treats transitive optional dependencies like any other transitive dependencies and therefore will fail if the dependencies doesn't exist.
For the installer, at least for now, we don't want the install to fail because of private jars.